+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 1.2.0
+
+- updated to OpenSimGauge 0.0.13
+- added throttle and train brake lever gauges
+- fixed `InvalidOperationException` errors
+- configure IP address and port with server options:
+
+```json
+{
+  "source": "DerailValley",
+  "sourceOptions": {
+    "ipAddress": "127.0.0.1",
+    "port": 1234
+  }
+}
+```
+
 # 1.1.0
 
 - updated to OpenSimGauge 0.0.10